+++ dpnp/backend/extensions/window/window_py.cpp
@@ -30,6 +30,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
+#include "bartlett.hpp"
 #include "blackman.hpp"
 #include "common.hpp"
 #include "hamming.hpp"
@@ -41,6 +42,7 @@ using window_ns::window_fn_ptr_t;
 
 namespace dpctl_td_ns = dpctl::tensor::type_dispatch;
 
+static window_fn_ptr_t bartlett_dispatch_vector[dpctl_td_ns::num_types];
 static window_fn_ptr_t blackman_dispatch_vector[dpctl_td_ns::num_types];
 static window_fn_ptr_t hamming_dispatch_vector[dpctl_td_ns::num_types];
 static window_fn_ptr_t hanning_dispatch_vector[dpctl_td_ns::num_types];
@@ -49,6 +51,21 @@ PYBIND11_MODULE(_window_impl, m)
 {
     using arrayT = dpctl::tensor::usm_ndarray;
     using event_vecT = std::vector<sycl::event>;
+
+    {
+        window_ns::init_window_dispatch_vectors<
+            window_ns::kernels::BartlettFactory>(bartlett_dispatch_vector);
+
+        auto bartlett_pyapi = [&](sycl::queue &exec_q, const arrayT &result,
+                                  const event_vecT &depends = {}) {
+            return window_ns::py_window(exec_q, result, depends,
+                                        bartlett_dispatch_vector);
+        };
+
+        m.def("_bartlett", bartlett_pyapi, "Call Bartlett kernel",
+              py::arg("sycl_queue"), py::arg("result"),
+              py::arg("depends") = py::list());
+    }
 
     {
         window_ns::init_window_dispatch_vectors<